[Ref] mpt/crypto: Silence signed-unsigned-comparison warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@15369 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ src/mpt/crypto/exception.hpp
@@ -69,7 +69,7 @@ public:
 inline void CheckNTSTATUS(NTSTATUS status) {
 	if (status >= 0) {
 		return;
-	} else if (status == STATUS_NO_MEMORY) {
+	} else if (static_cast<DWORD>(status) == STATUS_NO_MEMORY) {
 		mpt::throw_out_of_memory();
 	} else {
 		throw exception(status);
